# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de E-commerce Segura",
        "description": "API REST desarrollada con Laravel 12 para gestionar usuarios, productos, ordenes y pagos.",
        "contact": {
            "name": "Luis Monge"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Luis Monge"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "luis@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Iniciar sesion",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "luis@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesion exitoso"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Consultar el usuario autenticado",
                "operationId": "2baab1eeb5fd49516237a38da22e7a07",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado"
                    },
                    "401": {
                        "description": "Token no valido o no enviado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Cerrar sesion",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Sesion cerrada correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Ordenes"
                ],
                "summary": "Consultar el historial de compras",
                "description": "Devuelve las ordenes creadas por el usuario autenticado.",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "Historial de compras obtenido correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ordenes"
                ],
                "summary": "Crear una orden de compra",
                "description": "Crea una orden, calcula el total y descuenta el stock de los productos.",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrderItemInput"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "422": {
                        "description": "Datos invalidos, producto no disponible o stock insuficiente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Ordenes"
                ],
                "summary": "Consultar una orden por ID",
                "description": "Devuelve una orden solamente si pertenece al usuario autenticado.",
                "operationId": "3a2d78e76e9fc3f06bb879fb7d48b40d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden obtenida correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "403": {
                        "description": "La orden pertenece a otro usuario"
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItemInput": {
                "required": [
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token generado mediante Laravel Sanctum",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}